+++ euporie/render/__init__.py
@@ -26,6 +26,7 @@ from euporie.render.image import (
     img_ansi_timg_py,
     img_ansi_viu,
 )
+from euporie.render.latex import LatexRenderer, latex_flatlatex_py, latex_sympy_py
 from euporie.render.markdown import MarkdownRenderer, markdown_rich
 from euporie.render.mixin import (
     Base64Mixin,
@@ -56,6 +57,9 @@ __all__ = [
     "html_w3m",
     "MarkdownRenderer",
     "markdown_rich",
+    "LatexRenderer",
+    "latex_sympy_py",
+    "latex_flatlatex_py",
     "DataRendererMixin",
     "Base64Mixin",
     "Base64Mixin",