+++ euporie/render/base.py
@@ -45,7 +45,7 @@ class DataRenderer(metaclass=DataRendererMeta):
         self.width: "int" = width or 1
         self.height: "int" = height or 1
         self.graphic = graphic
-        self.bg_color: "str" = bg_color or self.app.term_info.bg_color or "#FFFFFF"
+        self.bg_color: "str" = bg_color or "#FFFFFF"
 
     def load(self, data: "Any") -> "None":
         """Function which performs setup tasks for the renderer.