+++ CHANGELOG.rst
@@ -4,6 +4,7 @@ Changlog
 Upcoming
 --------
 
+- Add ability to render LaTeX using ``sympy``
 - Add new terminal graphics system, which makes displaying using sixels / kitty graphics more reliable
 - Add support for LaTex equations in markdown using flatlatex
 - Markdown tables no longer expand to the full width of the display